+++ internal/cmd/root.go
@@ -23,8 +23,6 @@ var rootCmd = &cobra.Command{
 	Short: "Self-hosted virtual browser",
 	Long:  "A single Go binary that launches a browser instance (Chrome or Chromium) and streams it via WebRTC",
 	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-		setupLogging()
-
 		var err error
 		cfg, err = config.Load(cfgFile)
 		if err != nil {
@@ -36,6 +34,8 @@ var rootCmd = &cobra.Command{
 		}
 
 		cfg.ApplyEnv()
+
+		setupLogging()
 
 		level, err := zerolog.ParseLevel(cfg.Logging.Level)
 		if err != nil {
@@ -63,5 +63,17 @@ func init() {
 }
 
 func setupLogging() {
-	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
+	var output zerolog.ConsoleWriter
+	if cfg != nil && cfg.Logging.File != "" {
+		f, err := os.OpenFile(cfg.Logging.File, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+		if err != nil {
+			fmt.Fprintf(os.Stderr, "failed to open log file: %v\n", err)
+			output = zerolog.ConsoleWriter{Out: os.Stderr}
+		} else {
+			output = zerolog.ConsoleWriter{Out: f, NoColor: true}
+		}
+	} else {
+		output = zerolog.ConsoleWriter{Out: os.Stderr}
+	}
+	log.Logger = log.Output(output)
 }